requirements.txt
Django==1.8.4
MySQL-python==1.2.5
argparse==1.2.1
dj-database-url==0.3.0
dj-static==0.0.6
django-discover-runner==1.0
django-toolbelt==0.0.1
djangorestframework==3.3.1
gunicorn==19.3.0
pytz==2015.4
six==1.9.0
static3==0.6.1
wsgiref==0.1.2
